+++ src/pyvutils/asserts/__init__.py
@@ -3,6 +3,10 @@ from .inputs import (
     assert_input
 )
 
+from .files import (
+    assert_filename
+)
+
 from .currency import (
     assert_currency_code,
     assert_alphabetic_currency_code,