+++ src/pyvutils/asserts/files.py
@@ -1,0 +1,9 @@
+"""
+File Asserts
+"""
+
+import pathvalidate
+
+
+def assert_filename(filename: str) -> None:
+    pathvalidate.validate_filename(filename)